add chips: todo wrap them
