+++ src/data/ciscofulltime.js
@@ -4,6 +4,7 @@ export const ciscoFullTime = {
   subtitle: "Software Engineer",
   date: ["07/2023", "06/2025"],
   format: "write-up",
+  experience: true,
   summary: "Frontend and backend development, with a focus on AI/LLM applications. Developed Angular applications and worked on agentic AI projects.",
   content: [
     {
